Add a camera that automatically tries to focus on the graph's center.
I approach this by going through all displayed folder nodes and pick
those that define the minimum / maximum x and y values of the graph.

With these coordinates a rectangle is built and by bisecting its sides
we can get the mid point of the graph which is then tracked by the
camera.