Create a list of all commit-authors.
To get a clean list of the authors, the FileHandler removes the tag
which marks the line in the log as an "author-line" so we only get the
author's name.

Whenever a new commit is processed the author is stored in a list.

If LoGiVi finds a user-specified aliases.lua file it will use it to
replace the author names read from the log with aliases (this can be
helpful one author would be logged under two different names because of
a typo).