Add floating author objects that "shoot" beams at the files they modify.
The author objects spawn randomly across the screen and move in a random
direction until they hit a boundary and move in the opposite direction.

The link of each author to the files he edited is purged when a new
commit is loaded. This needs to be done or else each author would leave
the links from his last commit in the log and they would stay on the
screen forever.