Mark modified files with a color that fades away over time.
To fade the color we lerp between the default color for the file
extension (which it was assigned based on its extension) and the
"modified" color, which currently is a bright red.

After a small delay we mark the file as unmodified, reset the timer and
assign the default color (regardless of wether the lerping is done or
not).