User objects can be assigned an avatar.
This avatar can be set via the avatars.lua file in the save directory of
LoGiVi. Right now it is only possible to grab avatars from the internet,
by assigning an url that points directly to an image file.

LoGiVi will download this file to the computer of the user and then
load it as an image file.

If no avatar is assigned to an user, LoGiVi will use a default image.

There seems to be an issue with some urls currently, which I couldn't
fix yet.