Fix #2 - Catch error when trying to access an invalid file.
Instead of crashing LoGiVi will give out information about the file
and the used modifier. This seems to happen because due to merge commits
sometimes files don't show up as added (A-Modifier) in the log, which
means they won't be created by LoGiVi.

It also happens that a merge commit tries to delete (D-Modifier) a file
that has already been deleted.

There might be a way to fix this issue with the git log command (to make
it correctly take merge commits into account), but I have yet to find
it.