Print some information about the user's system to the console.
This includes the version of LoGiVi and the version of LÖVE the user
is running. Then the user's system is tested for compatibility with
certain features of the LÖVE framework (not all of them relevant for
the project yet). Last but not least information about the user's
graphics card is given.