Use email addresses to log authors.
Before storing a new author we check the config file to see if an alias
is associated with the specific email address. If it is, the alias will
be used and the nickname from the log is discarded.

If there isn't an alias, we check if there already is another name
stored for the email address. If there is, we will use the already
stored name.

If there isn't, we will use the name found in the log.