Use repulsion and attraction forces to form a basic layout.
All nodes are attracted to the center of the screen by a spring force.
The farther away they are from the center, the faster they will be
attracted.

Nodes repulse each other based on the distance between them and their
masses. The mass of each node is based on the amount of files they
carry. So nodes with a lot of files will have a bigger repulsion force.

Edges are hidden for now.