AuthorManager looks for avatars locally before trying to download them.
The path to an avatars on the local filesystem is set the same way as
an URL to an external image. LoGiVi simply checks if the string is a
valid path to a file and if it is not, it tries to use it as an url.

Also added a placehorder sprite which is used as a default avatar if
none has been specified.