Use string to store the human-readable date of each commit.
After transforming the date from a unix timestamp to a human-readable
date we directly store it as a string instead of returning a table,
which stores each value (mins, hours, ...) under a different key.

Tested on a repository with ~56.000 commits this saves about 20mb of
memory (which is about 25% of the total memory usage).