Fix #2 - Correctly remove edges to empty nodes.
First I removed the code for storing and handling edges separated from
the nodes. The graph now recursively calls all nodes, starting at the
root node and then lets each node draw the connections to its children.

The other important change was simplifying the code for deleting a node.
Since edges were handled independently it was a bit more complicated to
remove them. Now we just check if a node is empty (doesn't contain any
files) and doesn't have any children. In this case we remove it from
its parent and the graph.

Once it is removed from its parent, the edge to it will disappear too.