Fix #8 - Prevent crash when trying to modify a file which doesn't exist.
LoGiVi will log all occurrences where files should be added but already
exist or when they should either be removed or modified but don't
exist (anymore).

I don't quite know why this happens, but I suppose it's because a file
is "moved" (shows up as an addition and deletion operation in the log)
on one branch, while other branches still modify the original file. When
those branches are merged back, the changes are applied to the new file
(probably during the merge conflict resolve), but still show up in the
git log as if they modified the old file.