Close #10 - Allow the user to jump through time via the timeline
The user can now select a "timestep" in the timeline and LoGiVi will
jump to it using either the fast forward or the newly implemented fast
rewind function.

Just like the fast forward, the fast rewind function ignores modify
commits and quickly creates a correct representation of the graph just
by using the "add" and "delete" commits.

Also changed the look of the timeline. It is less intrusive and
hopefully also a bit more informative than before. When the user moves
the mouse into the clicking range of the timeline an indicator will
highlight the timestep the user will select by clicking.