Add an option to start visualization at the end of the log
The user can specify the play mode ('default' or 'rewind') in the
template. Default is just the normal running mode which starts at the
oldest commit and progresses to the newest.

Rewind starts at the last commit and progresses towards the first
commit instead.

To implement this, I had to find a way to quickly create the graph and
give it the shape it has in the last commit. This is where the
fast-forward function comes into play. It quickly processes all commits
between the current point in time and the target commit (in this case
the last one of the log). It filters out modification commits and only
cares about addition and deletion of nodes.