Add preliminary SelectionScreen which allows selection of different logs
The selection screen grabs the list of all logs found by the LogReader
and creates a button for each entry in the list. The UI is pretty
bare-bones - the buttons are just placed below each other and nothing
prevents them to go offscreen on a long list.

Once a button is pressed, the corresponding git log is set active and
the program switches to the MainScreen where the selected log is then
loaded and the visualization starts.