Fix #27 - Replace escape characters in the path to a repository
Paths loaded from the config file are now adjusted before being stored
and used for the creation of a log. All \-characters in the path will
be replaced with /-characters. (e.g.: C:\\Users\Robert\foo
> C:/Users/Robert/foo). Before this, the \ character would break the
path since it is used as an escape character in Lua.

This also seems to work on Windows without problems.