Fix controls
Transforms the keyconstants which love.keypressed receives as parameters
to numbers (if possible) and the ' ' character to 'space'. This is done
so they correspond to the keybindings read from the config file.

Maybe it would be "cleaner" to fix this in the config reader, but for
now this should work.