Fix #5 - Improve author movement by applying just a spring force
Rewrote most of the movement related code in the author class. Similar
to the nodes in the graph it now uses a spring force to attract authors
towards nodes. Authors aren't moving away too far from the graph anymore
(except for in some rare cases). The movement is more fluent than before
and authors move towards the parts of the graph they currently work on.