Improve code for removal of dead nodes
The function now only takes the node as a parameter (the path is fetched
via the getter of the node). Instead of calling this function after a
file has been removed (when a commit has been processed) we call it each
update cycle. This should speed things up when a lot of files are
removed in a single commit.