Close #50 - Display warning message when git isn't available
When the LoGiVi can't find a git installation on the user's system it
will display a MessageBox and notify the user that logs won't be created
automatically.

It will offer an option to open the online instructions or to just
proceed to the selection screen, where the user still can view logs he
created manually.

Note: There is a small issue with assigning both the enter and escape
keys to the button index 1. This seems to be an issue with SDL on OSX
(https://bitbucket.org/rude/love/issues/1102).