Implement Graphoon library
All the code responsible for calculating the force directed graph layout
has been moved to its own library. The library has a cleaner way of
storing nodes and edges. Instead of using nested nodes (children belong
to their parents) all nodes are on the same level so to speak. Nodes are
independent until they are connected by an Edge.

All in all this system makes it easier to maintain the code since the
physical calculations are handled "outside" of LoGiVi now. Besides that,
all the code for drawing the graph can now be found in the Graph class
itself.