Completely rewrite the SelectionScreen menu
Instead of using a ui we use the graph generated by the loading screen.

Each node represents a git repository. The user can click on a node to
start watching the visualisation. If the "root" node is clicked LoGiVi
will open the save folder containing all the logs and the settings file.